+++ package.json
@@ -51,7 +51,7 @@
     "cross-env": "^7.0.3",
     "css-loader": "^5.2.6",
     "cssnano": "^5.0.2",
-    "enonic-types": "^0.2.6",
+    "enonic-types": "^0.2.11",
     "eslint": "^6.8.0",
     "eslint-config-google": "^0.14.0",
     "eslint-plugin-react": "^7.24.0",
