+++ src/main/resources/tasks/cleanupPublishDataset/cleanupPublishDataset-config.ts
@@ -1,0 +1,22 @@
+// WARNING: This file was automatically generated by "no.item.xp.codegen". You may lose your changes if you edit it.
+export interface CleanupPublishDatasetConfig {
+  /**
+   * JobId
+   */
+  jobId: string;
+
+  /**
+   * StatisticsContentId
+   */
+  statisticsContentId: string;
+
+  /**
+   * StatisticsId
+   */
+  statisticsId: string;
+
+  /**
+   * PublicationItem
+   */
+  publicationItem: string;
+}
