+++ src/main/resources/tasks/publishDataset/publishDataset-config.ts
@@ -1,0 +1,22 @@
+// WARNING: This file was automatically generated by "no.item.xp.codegen". You may lose your changes if you edit it.
+export interface PublishDatasetConfig {
+  /**
+   * JobId
+   */
+  jobId: string;
+
+  /**
+   * StatisticsContentId
+   */
+  statisticsContentId: string;
+
+  /**
+   * StatisticsId
+   */
+  statisticsId: string;
+
+  /**
+   * PublicationItem
+   */
+  publicationItem: string;
+}
