+++ src/main/resources/tasks/publishDataset/publishDataset.xml
@@ -1,0 +1,21 @@
+<task>
+  <description>Publish dataset</description>
+  <form>
+    <input type="TextLine" name="jobId">
+      <label>JobId</label>
+      <occurrences minimum="1" maximum="1"/>
+    </input>
+    <input type="TextLine" name="statisticsContentId">
+      <label>StatisticsContentId</label>
+      <occurrences minimum="1" maximum="1"/>
+    </input>
+    <input type="TextLine" name="statisticsId">
+      <label>StatisticsId</label>
+      <occurrences minimum="1" maximum="1"/>
+    </input>
+    <input type="TextArea" name="publicationItem">
+      <label>PublicationItem</label>
+      <occurrences minimum="1" maximum="1"/>
+    </input>
+  </form>
+</task>
