+++ src/main/resources/tasks/publishJob/publishJob.xml
@@ -1,0 +1,5 @@
+<task>
+  <description>Publish job</description>
+  <form>
+  </form>
+</task>
